requirements.txt
pytest~=8.3.1
selenium~=4.23.0
Faker~=26.0.0
randstr~=0.1.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
